Remove unused plugins and make changes in opus_codec attack
