+++ .env.example
@@ -17,6 +17,4 @@ NEURAL_VOCODER_PORT=10004
 SPEECH_ENHANCEMENT_PORT1=10005
 SPEECH_ENHANCEMENT_PORT2=10006
 NETWORK_TRANSMISSION_PORT=10020
-NETWORK_TRANSMISSION_2_PORT=10021
-NETWORK_TRANSMISSION_3_PORT=10022
 OPUS_CODEC_PORT=10023
