+++ docker-compose.yml
@@ -120,28 +120,6 @@ services:
     cap_add:
       - NET_ADMIN
 
-  network_transmission_2:
-    build:
-      context: .
-      dockerfile: src/plugins/attacks/network_transmission_2/Dockerfile
-    ports:
-      - "${NETWORK_TRANSMISSION_2_PORT}:${NETWORK_TRANSMISSION_2_PORT}"
-    environment:
-      - APP_PORT=${NETWORK_TRANSMISSION_2_PORT}
-    cap_add:
-      - NET_ADMIN
-
-  network_transmission_3:
-    build:
-      context: .
-      dockerfile: src/plugins/attacks/network_transmission_3/Dockerfile
-    ports:
-      - "${NETWORK_TRANSMISSION_3_PORT}:${NETWORK_TRANSMISSION_3_PORT}"
-    environment:
-      - APP_PORT=${NETWORK_TRANSMISSION_3_PORT}
-    cap_add:
-      - NET_ADMIN
-
   opus_codec:
     build:
       context: .
