+++ src/plugins/attacks/network_transmission/Dockerfile
@@ -1,20 +1,18 @@
 FROM python:3.10-slim
 
-# Default port; overridden at runtime by docker-compose (APP_PORT from .env)
 ENV APP_PORT=10020
 ENV HOST="0.0.0.0"
 
-# Install system dependencies
-# - iproute2: for tc (traffic control) with netem
-# - opus-tools: for opusenc/opusdec
-# - libopus-dev: opus library
+# - iproute2: tc netem
+# - libopus-dev / libopus0: opuslib needs the Opus C library
 # - libsndfile1: for soundfile
-# - swig, build-essential, python3-dev: for building python-webrtc-audio-processing
-# - git: for initializing submodules
+# - swig, build-essential, python3-dev, libc++-dev: required to build
+#   python-webrtc-audio-processing from source
+# - git: to clone the build sources
 RUN apt-get update && apt-get install -y --no-install-recommends \
     iproute2 \
-    opus-tools \
     libopus-dev \
+    libopus0 \
     libsndfile1 \
     swig \
     build-essential \
@@ -29,10 +27,10 @@ WORKDIR /app
 COPY src/plugins/attacks/network_transmission/requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
-# Build and install python-webrtc-audio-processing (clone from fork with cross-platform support)
+# Build python-webrtc-audio-processing from the same fork v1 uses, with
+# the pinned upstream commit known to compile cleanly.
 RUN git clone https://github.com/nesovicelena/python-webrtc-audio-processing.git /tmp/python-webrtc-audio-processing && \
     cd /tmp/python-webrtc-audio-processing && \
-    # Clone webrtc-audio-processing submodule (pinned to compatible commit)
     git clone https://gitlab.freedesktop.org/pulseaudio/webrtc-audio-processing.git webrtc-audio-processing && \
     cd webrtc-audio-processing && git checkout ee8cfef49b8417c2f0ba65a249d2ee8c360d19ab && cd .. && \
     python setup.py build && \
@@ -41,9 +39,6 @@ RUN git clone https://github.com/nesovicelena/python-webrtc-audio-processing.git
 
 COPY src/plugins/attacks/network_transmission .
 
-# Expose the default port
 EXPOSE 10020
 
-# Run the FastAPI server using Uvicorn
-# Note: Container must be run with --cap-add=NET_ADMIN for tc to work
 CMD uvicorn app:app --host=$HOST --port=$APP_PORT
