+++ src/plugins/attacks/network_transmission/requirements.txt
@@ -4,3 +4,5 @@ uvicorn
 numpy
 soundfile
 librosa
+opuslib
+pyloudnorm
