--- src/plugins/attacks/network_transmission_2/requirements.txt
@@ -1,8 +1,0 @@
-pydantic
-fastapi
-uvicorn
-numpy
-soundfile
-librosa
-opuslib
-noisereduce
