--- src/plugins/attacks/network_transmission_3/requirements.txt
@@ -1,8 +1,0 @@
-pydantic
-fastapi
-uvicorn
-numpy
-soundfile
-librosa
-opuslib
-pyloudnorm
