+++ src/plugins/attacks/opus_codec/config.json
@@ -1,5 +1,4 @@
 {
     "bitrate_opus_codec": 16,
-    "framesize_opus_codec": 20,
-    "sampling_rate_opus_codec": 16000
+    "framesize_opus_codec": 20
 }
